task/jenkins 39478 popover refactor (#109)
* [JENKINS-39478] refactor core Popover logic into more generic "FloatingElement"

* [JENKINS-39478] cleanup

* [JENKINS-39478] cleanup

* [JENKINS-39478] fix flow error

* [JENKINS-39478] fix flow error
