+++ less/components.less
@@ -1,4 +1,5 @@
 @import "components/empty-state";
+@import "components/floating-element";
 @import "components/result-status";
 @import "components/expandable-path";
 @import "components/favorite";
