+++ less/components/floating-element.less
@@ -1,0 +1,15 @@
+.FloatingElement {
+    position: fixed;
+    left: 0;
+    top: 0;
+    width: 0;
+    height: 0;
+    z-index: 100;
+}
+
+.FloatingElement-wrapper {
+    position: fixed;
+    left: 0;
+    top: 0;
+}
+
