+++ less/components/popover.less
@@ -1,11 +1,3 @@
-.Popover {
-    position: fixed;
-    left: 0;
-    top: 0;
-    width: 0;
-    height: 0;
-}
-
 .Popover-modalScreen {
     position: fixed;
     top: 0;
@@ -18,10 +10,6 @@
 }
 
 .Popover-wrapper {
-    position: fixed;
-    z-index: 100;
-    left: 0;
-    top: 0;
     filter: drop-shadow(0px 2px 4px rgba(0, 0, 0, 0.5));
 }
