+++ src/js/components/index.js
@@ -34,3 +34,6 @@ export {TimeDuration} from './TimeDuration';
 export { Progress } from './Progress';
 export { Popover } from './Popover';
 export {ExpandablePath} from './ExpandablePath';
+
+import FloatingElement from './FloatingElement';
+export {FloatingElement};
